Start moving search into the top bar
